#101: Rename StaticString.h to StaticArray.h
