+++ tok3n.vcxproj
@@ -247,7 +247,7 @@
     <ClInclude Include="include\k3\tok3n\types\Span.h" />
     <ClInclude Include="include\k3\tok3n\types\ParserType.h" />
     <ClInclude Include="include\k3\tok3n\types\Tags.h" />
-    <ClInclude Include="include\k3\tok3n\types\StaticString.h" />
+    <ClInclude Include="include\k3\tok3n\types\StaticArray.h" />
     <ClInclude Include="include\k3\tok3n\modifiers\multi.h" />
     <ClInclude Include="include\k3\tok3n\detail\type_list.h" />
     <ClInclude Include="include\k3\tok3n\parsers\basic\_fwd.h" />
